+++ packages/collector-obd/src/collector_obd/protocol/__init__.py
@@ -1,0 +1,36 @@
+"""OBD protocol models and parsing helpers."""
+
+from collector_obd.protocol.commands import (
+    AT_ECHO_OFF,
+    AT_HEADERS_OFF,
+    AT_LINEFEEDS_OFF,
+    AT_PROTOCOL_AUTO,
+    AT_SPACES_OFF,
+    ATZ,
+    CONTROL_MODULE_VOLTAGE,
+    COOLANT_TEMPERATURE,
+    RPM,
+    initialization_commands,
+    pid_commands,
+)
+from collector_obd.protocol.models import ObdCommand, ObdPid, ObdResponse, ObdSession
+from collector_obd.protocol.parser import parse_pid_response
+
+__all__ = [
+    "AT_ECHO_OFF",
+    "AT_HEADERS_OFF",
+    "AT_LINEFEEDS_OFF",
+    "AT_PROTOCOL_AUTO",
+    "AT_SPACES_OFF",
+    "ATZ",
+    "CONTROL_MODULE_VOLTAGE",
+    "COOLANT_TEMPERATURE",
+    "RPM",
+    "ObdCommand",
+    "ObdPid",
+    "ObdResponse",
+    "ObdSession",
+    "initialization_commands",
+    "parse_pid_response",
+    "pid_commands",
+]